+++ src/packages/user/section/sidebar-app/index.ts
@@ -1,1 +1,1 @@
-export * from './constants.js';
+export * from '../constants.js';
