+++ src/packages/user/user-group/manifests.ts
@@ -7,9 +7,9 @@ import { manifests as repositoryManifests } from './repository/manifests.js';
 import { manifests as sectionViewManifests } from './section-view/manifests.js';
 import { manifests as workspaceManifests } from './workspace/manifests.js';
 
-import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestTypes, UmbBackofficeManifestKind } from '@umbraco-cms/backoffice/extension-registry';
 
-export const manifests: Array<ManifestTypes> = [
+export const manifests: Array<ManifestTypes | UmbBackofficeManifestKind> = [
 	...collectionManifests,
 	...entityActionManifests,
 	...entityBulkActionManifests,
