+++ src/packages/user/user/workspace/user/components/user-workspace-access/user-workspace-access.element.ts
@@ -1,5 +1,5 @@
 import { UMB_USER_WORKSPACE_CONTEXT } from '../../user-workspace.context-token.js';
-import type { UmbUserStartNodesModel } from '../../../types.js';
+import type { UmbUserStartNodesModel } from '../../../../types.js';
 import { html, customElement, state, css } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
