+++ src/packages/user/user/workspace/user/components/user-workspace-info/user-workspace-info.element.ts
@@ -1,8 +1,8 @@
-import type { UmbUserDisplayStatus } from '../../../utils.js';
-import { TimeFormatOptions, getDisplayStateFromUserStatus } from '../../../utils.js';
+import type { UmbUserDisplayStatus } from '../../../../utils.js';
+import { TimeFormatOptions, getDisplayStateFromUserStatus } from '../../../../utils.js';
 import { UMB_USER_WORKSPACE_CONTEXT } from '../../user-workspace.context-token.js';
-import type { UmbUserDetailModel } from '../../../types.js';
-import { UmbUserKind } from '../../../utils/index.js';
+import type { UmbUserDetailModel } from '../../../../types.js';
+import { UmbUserKind } from '../../../../utils/index.js';
 import { html, customElement, state, css, repeat, ifDefined, nothing } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
