+++ src/packages/user/user/workspace/user/manifests.ts
@@ -1,0 +1,42 @@
+import { UMB_USER_ENTITY_TYPE } from '../../entity.js';
+import { UMB_USER_WORKSPACE_ALIAS } from './constants.js';
+import { UmbSubmitWorkspaceAction } from '@umbraco-cms/backoffice/workspace';
+import type {
+	ManifestWorkspaces,
+	ManifestWorkspaceActions,
+	ManifestTypes,
+} from '@umbraco-cms/backoffice/extension-registry';
+
+const workspace: ManifestWorkspaces = {
+	type: 'workspace',
+	kind: 'routable',
+	alias: UMB_USER_WORKSPACE_ALIAS,
+	name: 'User Workspace',
+	api: () => import('./user-workspace.context.js'),
+	meta: {
+		entityType: UMB_USER_ENTITY_TYPE,
+	},
+};
+
+const workspaceActions: Array<ManifestWorkspaceActions> = [
+	{
+		type: 'workspaceAction',
+		kind: 'default',
+		alias: 'Umb.WorkspaceAction.User.Save',
+		name: 'Save User Workspace Action',
+		api: UmbSubmitWorkspaceAction,
+		meta: {
+			label: '#buttons_save',
+			look: 'primary',
+			color: 'positive',
+		},
+		conditions: [
+			{
+				alias: 'Umb.Condition.WorkspaceAlias',
+				match: workspace.alias,
+			},
+		],
+	},
+];
+
+export const manifests: Array<ManifestTypes> = [workspace, ...workspaceActions];
