+++ src/packages/user/user/workspace/user/user-workspace.context-token.ts
@@ -1,4 +1,4 @@
-import { UMB_USER_ENTITY_TYPE } from '../entity.js';
+import { UMB_USER_ENTITY_TYPE } from '../../entity.js';
 import type { UmbUserWorkspaceContext } from './user-workspace.context.js';
 import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 import type { UmbSubmittableWorkspaceContext } from '@umbraco-cms/backoffice/workspace';
