+++ src/packages/user/user/workspace/user/user-workspace.context.ts
@@ -1,8 +1,8 @@
-import type { UmbUserDetailModel, UmbUserStartNodesModel, UmbUserStateEnum } from '../types.js';
-import { UMB_USER_ENTITY_TYPE } from '../entity.js';
-import { UmbUserDetailRepository } from '../repository/index.js';
-import { UmbUserAvatarRepository } from '../repository/avatar/index.js';
-import { UmbUserConfigRepository } from '../repository/config/index.js';
+import type { UmbUserDetailModel, UmbUserStartNodesModel, UmbUserStateEnum } from '../../types.js';
+import { UMB_USER_ENTITY_TYPE } from '../../entity.js';
+import { UmbUserDetailRepository } from '../../repository/index.js';
+import { UmbUserAvatarRepository } from '../../repository/avatar/index.js';
+import { UmbUserConfigRepository } from '../../repository/config/index.js';
 import { UMB_USER_WORKSPACE_ALIAS } from './constants.js';
 import { UmbUserWorkspaceEditorElement } from './user-workspace-editor.element.js';
 import type { UmbSubmittableWorkspaceContext } from '@umbraco-cms/backoffice/workspace';
